refactor(ui): Rename other curation related components
Some other components were also renamed to match the new naming.
