+++ apps/clearance_ui/src/components/license_conclusions/ConclusionForm.tsx
@@ -31,9 +31,9 @@ import {
     TooltipTrigger,
 } from "@/components/ui/tooltip";
 import { useToast } from "@/components/ui/use-toast";
-import CurationDB from "@/components/license_conclusions/CurationDB";
-import CurationLicense from "@/components/license_conclusions/CurationLicense";
-import CurationSPDX from "@/components/license_conclusions/CurationSPDX";
+import CurationDB from "@/components/license_conclusions/ConclusionDB";
+import CurationLicense from "@/components/license_conclusions/ConclusionLicense";
+import CurationSPDX from "@/components/license_conclusions/ConclusionSPDX";
 import { toPathPurl } from "@/helpers/pathParamHelpers";
 import { cn } from "@/lib/utils";
